Fix: unable to type in screen on big-endian systems
Thanks to Petr Sumbera <petr.sumbera@oracle.com>
and Dan Horák <dan@danny.cz>

bug#67607

+++ src/telnet.h
@@ -9,7 +9,7 @@
 int TelOpenAndConnect(Window *);
 int TelIsline(Window *);
 void TelProcessLine(char **, size_t *);
-int DoTelnet(char *, size_t *, int);
+int DoTelnet(char *, int *, int);
 int TelIn(Window *, char *, size_t, int);
 void TelBreak(Window *);
 void TelWindowSize(Window *);
